Implemented calculation of MAX_BTN and change initial configuration generation
- Implemented code that calculates maximum possible integer value of BTN taking into account iterations, sampling rate and cutoff value.
- Changed code for generation of first configuration to calculate magnitude from 3kT being the trapping potential for each position component.

+++ src/analysis.py
@@ -67,7 +67,7 @@ def plot_snapshot(positions: np.ndarray[np.float64], iteration: int) -> None:
 
 def plot_error(errors: np.ndarray[np.float64]) -> None:
     _, ax = plt.subplots()
-    ax.plot(np.linspace(1, len(errors), len(errors)), errors)
+    ax.plot(np.linspace(1, len(errors), len(errors), dtype="int64"), errors)
     ax.set(xlabel="BTN", ylabel="Standard Error")
     plt.show()
